refactor: 저장 스냅샷과 mapper를 jdbc 어댑터 내부로 이동한다
