+++ src/main/java/Application.java
@@ -4,9 +4,9 @@ import controller.GameController;
 import infra.jdbc.JdbcConnectionManager;
 import infra.jdbc.JdbcGameRepository;
 import infra.jdbc.SchemaInitializer;
+import infra.jdbc.SavedGameReadMapper;
+import infra.jdbc.SavedGameWriteMapper;
 import java.time.Clock;
-import repository.SavedGameReadMapper;
-import repository.SavedGameWriteMapper;
 import view.GamePresenter;
 import view.InputView;
 import view.OutputView;
