+++ src/test/java/infra/jdbc/SavedGameReadMapperTest.java
@@ -1,4 +1,4 @@
-package repository;
+package infra.jdbc;
 
 import static org.assertj.core.api.Assertions.assertThat;
