+++ src/test/java/infra/jdbc/SavedGameWriteMapperTest.java
@@ -1,4 +1,4 @@
-package repository;
+package infra.jdbc;
 
 import static org.assertj.core.api.Assertions.assertThat;
